+++ packages/web-components/fast-foundation/src/design-token/design-token.ts
@@ -86,7 +86,9 @@ export interface CSSDesignToken<
         | null
         | Array<any>
         | symbol
-        | { createCSS?(): string }
+        | ({
+              createCSS?(): string;
+          } & Record<PropertyKey, any>)
 > extends DesignToken<T>, CSSDirective {
     /**
      * The {@link (DesignToken:interface)} formatted as a CSS custom property if the token is
@@ -192,10 +194,10 @@ class DesignTokenImpl<T extends { createCSS?(): string }> extends CSSDirective
     ): this {
         this._appliedTo.add(element);
         if (value instanceof DesignTokenImpl) {
-            const _value = value;
-            value = ((_element: HTMLElement) =>
-                DesignTokenNode.for<T>(_value, _element)
-                    .value) as DerivedDesignTokenValue<T>;
+            const tokenValue = value;
+
+            value = ((target: HTMLElement) =>
+                tokenValue.getValueFor(target)) as DerivedDesignTokenValue<T>;
         }
         DesignTokenNode.for<T>(this, element).set(value);
         [